+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 2.0.0
 
-- __Breaking__: `FieldScope.field`, `FieldCompareQuery.field` and `FieldCompareQuery.operator` are now a `TextQuery`.
-- __Breaking__: `field:<exp>` and `field <operator> <exp>`: `<exp>` can now be an empty string.
-- Added `startIndex` and `endIndex` to `Query`.
+**Breaking changes**
+- `FieldScope.field`, `FieldCompareQuery.field` and `FieldCompareQuery.operator` are now a `TextQuery`.
+- `field:<exp>` and `field <operator> <exp>`: `<exp>` can now be an empty string.
+
+**Updated**
+- Added `SourcePosition` to `Query`, storing the `start` and `end` index of the matching input.
 
 ## 1.6.0